Extract colour and styles into separate presentation modules.
